app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
databricks-sdk>=0.20.0
httpx>=0.25.0
python-multipart>=0.0.9
pydantic>=2.0.0
openpyxl>=3.1.0
